Move test helpers into a shared file
